feature: subscription override field (#18)
<!-- Thank you for submitting a Pull Request. Please fill out the
template below.-->
## Overview/Summary

Adds an `azure_subscription_id` variable that defaults to the connected
subscription if left blank. This can be used to target a subscription
other than the one connected to for the bootstrap resources.

A future iteration could update this field to use the
`subscription_id_management` field in the starter module, but would
require updates to the ALZ PowerShell module to achieve that in a robust
way.

## This PR fixes/adds/changes/removes

1. #15 

### Breaking Changes

None. Defaults to the connected subscription if none is supplied, which
is the existing behaviour.

## Testing Evidence

Tested with this version of ALZ PowerShell module:
https://github.com/Azure/ALZ-PowerShell-Module/pull/77

## As part of this Pull Request I have

- [x] Checked for duplicate [Pull
Requests](https://github.com/Azure/alz-terraform-accelerator/pulls)
- [x] Associated it with relevant
[issues](https://github.com/Azure/alz-terraform-accelerator/issues), for
tracking and closure.
- [x] Ensured my code/branch is up-to-date with the latest changes in
the `main`
[branch](https://github.com/Azure/alz-terraform-accelerator/tree/main)
- [x] Performed testing and provided evidence.
- [x] Updated relevant and associated documentation.