bug: fixes for github and resource providers (#32)
<!-- Thank you for submitting a Pull Request. Please fill out the
template below.-->
## Overview/Summary

A few bugs have been picked up since the plan and apply perms were
split:

- The registering of resource providers on a brand new sub is not
possible with the plan permissions
- The GitHub PR comment does not work for very large plans
- An error has been reported for another permission:

```
403 Code="AuthorizationFailed" Message="The client '83a5fcf5-b664-4278-84b3-d6c5a1ff608a' with object id '83a5fcf5-b664-4278-84b3-d6c5a1ff608a' does not have authorization to perform action 'Microsoft.Resources/deployments/exportTemplate/action' over scope '/subscriptions/6f734e46-d97b-495f-9be8-f600638f3808' or the scope is invalid
```

This permissions issue is related to the telemetry settings for the CAF
module. Given this is only for internal use, the solution is to disable
the telemetry.

## This PR fixes/adds/changes/removes

1. #33

### Breaking Changes

1. None

## Testing Evidence

The E2E tests will be run.

I also ran a manual test of deploying the basic template from GitHub and
then running another plan / apply. This was successful, where it failed
prior to these fixes.

## As part of this Pull Request I have

- [x] Checked for duplicate [Pull
Requests](https://github.com/Azure/alz-terraform-accelerator/pulls)
- [x] Associated it with relevant
[issues](https://github.com/Azure/alz-terraform-accelerator/issues), for
tracking and closure.
- [x] Ensured my code/branch is up-to-date with the latest changes in
the `main`
[branch](https://github.com/Azure/alz-terraform-accelerator/tree/main)
- [x] Performed testing and provided evidence.
- [x] Updated relevant and associated documentation.