docs: update url in example readmes (#236)
<!-- Thank you for submitting a Pull Request. Please fill out the
template below.-->
## Overview/Summary

For some reason the URL to the Azure Landing Zones documentation site in
all of the example `README` files points to a financial institution and
not a Microsoft owned website. For simplicity I have just updated the
urls to use
https://azure.github.io/Azure-Landing-Zones/accelerator/startermodules/terraform-platform-landing-zone/scenarios/
where an overview of all the scenarios are listed

## This PR fixes/adds/changes/removes

1. Faulty URLs in `README` files

## Testing Evidence

not applicable. This pr doesn't touch any code

## As part of this Pull Request I have

- [x] Checked for duplicate [Pull
Requests](https://github.com/Azure/alz-terraform-accelerator/pulls)
- [ ] Associated it with relevant
[issues](https://github.com/Azure/alz-terraform-accelerator/issues), for
tracking and closure.
- [x] Ensured my code/branch is up-to-date with the latest changes in
the `main`
[branch](https://github.com/Azure/alz-terraform-accelerator/tree/main)
- [ ] Performed testing and provided evidence.
- [x] Updated relevant and associated documentation.