fix: remove byo ddos plan as it should not be in the config by default (#238)
<!-- Thank you for submitting a Pull Request. Please fill out the
template below.-->
## Overview/Summary

The setting for BYO DDOS protection plan was set to the DDOS plan being
created, so it worked by luck. It should not be set by default though,
it should only be set when the DDOS plan is created outside of this
module.

## This PR fixes/adds/changes/removes

N/A

### Breaking Changes

None

## Testing Evidence

Please provide any testing evidence to show that your Pull Request
works/fixes as described and planned (include screenshots, if
appropriate).

## As part of this Pull Request I have

- [x] Checked for duplicate [Pull
Requests](https://github.com/Azure/alz-terraform-accelerator/pulls)
- [x] Associated it with relevant
[issues](https://github.com/Azure/alz-terraform-accelerator/issues), for
tracking and closure.
- [x] Ensured my code/branch is up-to-date with the latest changes in
the `main`
[branch](https://github.com/Azure/alz-terraform-accelerator/tree/main)
- [x] Performed testing and provided evidence.
- [x] Updated relevant and associated documentation.